moved class with HTTP header names to vocabulary package
